Migrated from double to decimal

+++ Business/Models/AccountStatus.cs
@@ -4,7 +4,7 @@ namespace Business.Models
 {
     public class AccountStatus
     {
-        public AccountStatus(int accountStatusId, int accountId, DateTime date, double amount)
+        public AccountStatus(int accountStatusId, int accountId, DateTime date, decimal amount)
         {
             AccountStatusId = accountStatusId;
             AccountId = accountId;
@@ -15,6 +15,6 @@ namespace Business.Models
         public int AccountStatusId { get; private set; }
         public int AccountId { get; private set; }
         public DateTime Date { get; private set; }
-        public double Amount { get; private set; }
+        public decimal Amount { get; private set; }
     }
 }
