+++ Legacy/Models/LegacyPosting.cs
@@ -5,11 +5,11 @@ namespace Legacy.Models
     public class LegacyPosting
     {
         public int Posting_id { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public DateTime Date { get; set; }
         public required string Description { get; set; }
-        public double? Latitude { get; set; }
-        public double? Longitude { get; set; }
-        public double? Accuracy { get; set; }
+        public decimal? Latitude { get; set; }
+        public decimal? Longitude { get; set; }
+        public decimal? Accuracy { get; set; }
     }
 }
