+++ Legacy/Repositories/ILegacyAccountStatusQueryRepository.cs
@@ -5,6 +5,6 @@ namespace Legacy.Repositories
     public interface ILegacyAccountStatusQueryRepository
     {
         MonthlyAccountCategorySums GetAccumulatedCategorySums(int userId);
-        (double StartSum, double EndSum) StartAndEndOfMonthAccountStatusSum(int userId, MonthAndYear monthAndYear);
+        (decimal StartSum, decimal EndSum) StartAndEndOfMonthAccountStatusSum(int userId, MonthAndYear monthAndYear);
     }
 }
