+++ Legacy/Repositories/ILegacyPostingQueryRepository.cs
@@ -7,8 +7,8 @@ namespace Legacy.Repositories
     public interface ILegacyPostingQueryRepository
     {
         List<LegacyMonthlyTypeSumWithColorAndName> GetMonthlyStatus(int userId, int year, int month);
-        (object, Dictionary<int, Dictionary<int, Dictionary<int, Dictionary<int, double>>>>) Summary(int userId);
-        double GetMonthlyChange(int userId, MonthAndYear monthAndYear);
+        (object, Dictionary<int, Dictionary<int, Dictionary<int, Dictionary<int, decimal>>>>) Summary(int userId);
+        decimal GetMonthlyChange(int userId, MonthAndYear monthAndYear);
         Dictionary<MonthAndYear, IncomeAndExpenses> GetMonthlyIncomeAndExpenses(int userId);
     }
 }
